Move upload functionality to user service
