+++ services/conversion/main.py
@@ -1,0 +1,6 @@
+
+def _main():
+    pass
+
+if __name__ == '__main__':
+    _main()
