+++ services/conversion/requirements.txt
@@ -1,5 +1,3 @@
-Flask~=3.0.3
-flask_http_middleware~=0.4.2
 pymongo~=4.7.2
 docx~=0.2.4
 docx2txt~=0.8
