--- services/conversion/templates/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>File Converter</title>
-</head>
-<body>
-<h1>File Converter</h1>
-<form method="POST" enctype="multipart/form-data">
-    <input type="file" name="file" required>
-    <button type="submit">Convert</button>
-</form>
-</body>
-</html>
