+++ services/user_service/Dockerfile
@@ -14,4 +14,4 @@ COPY . .
 EXPOSE 5000
 
 # Run the Flask application with opentelemetry instrumentation
-CMD ["opentelemetry-instrument", "python", "main.py"]
+CMD ["opentelemetry-instrument", "flask", "run", "--host=0.0.0.0"]
