+++ services/user_service/templates/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>File Uploader</title>
+</head>
+<body>
+<h1>File Converter</h1>
+<form action="/upload" enctype="multipart/form-data" method="POST">
+    <input name="file" required type="file">
+    <button type="submit">Upload</button>
+</form>
+</body>
+</html>
